requirements.txt
PIL
pygame
seaborn
matplotlib
numpy
gym
tensorflow-gpu==2.0

